+++ src/NoCache.php
@@ -11,8 +11,10 @@ use Illuminate\Http\Request;
 class NoCache
 {
     /**
-     * @param Request $request HTTP Request
-     * @param Closure $next Next middleware handler
+     * Handle an incoming request.
+     *
+     * @param  Request $request HTTP Request to process
+     * @param  Closure $next Next handler in chain
      * @return mixed
      */
     public function handle(Request $request, Closure $next)
@@ -21,7 +23,7 @@ class NoCache
 
         $response->header('Cache-Control', 'no-cache, no-store, must-revalidate');
         $response->header('Pragma', 'no-cache');
-        $response->header('Expires', '0');
+        $response->header('Expires', 'Fri, 01 Jan 1990 00:00:00 GMT');
 
         return $response;
     }