+++ CHANGES.md
@@ -1,5 +1,17 @@
 # Changes History
 
+1.0.6
+-----
+Add RequestChecker utility to check if request is secure and contains header value or not
+
+1.0.5
+-----
+Add AjaxOnly middleware
+
+1.0.4
+-----
+Update dependencies versions
+
 1.0.3
 -----
 - Update dependencies versions